Type the Redux store
